requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.7.0

# Error Monitoring
sentry-sdk[fastapi]==2.19.2

# Database Migrations (optional - for schema version control)
alembic==1.14.0

supabase==2.11.0
postgrest==0.19.0
gotrue==2.11.2
httpx==0.27.2
httpcore==1.0.6

psycopg2-binary==2.9.9

groq==0.4.2
langchain==0.1.11
langchain-groq==0.0.1
sentence-transformers==2.3.1


playwright==1.42.0

python-jose[cryptography]==3.3.0
# Update this line in your requirements.txt
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

python-dotenv==1.0.1
aiofiles==23.2.1

pytest==8.0.2
pytest-asyncio==0.23.5

# Authentication - FIXED
pyjwt==2.8.0
pydantic[email]
pydantic_settings==2.13.1

passlib[bcrypt]>=1.7.4

resend==2.23.0
sib-api-v3-sdk==7.6.0

loguru==0.7.3
slowapi==0.1.9
redis==7.4.0
upstash-redis==1.7.0
